Reorganize rust request structure
Move code related to request logic into separate folder.
