+++ src/lib.rs
@@ -4,7 +4,7 @@ extern crate napi_derive;
 // Link other files
 pub mod auth;
 pub mod options;
-pub mod request;
+pub mod requests;
 pub mod result;
 pub mod session;
 pub mod tests;
